Version final proyecto para entregar

+++ src/main/resources/coderBD.sql
@@ -2,7 +2,7 @@ create database coder;
 use coder;
 
 create table clients (
-`id` int primary key auto_increment,
+`id` bigint primary key auto_increment,
 `dni` varchar (10) unique not null,
 `name` varchar(50)  not null,
 `lastname` varchar(50) not null,
@@ -10,7 +10,7 @@ create table clients (
 );
 
 create table products (
-`id` int primary key auto_increment,
+`id` bigint primary key auto_increment,
 `sku` varchar(100),
 `description` varchar(200),
 `buy_price` double,
@@ -20,17 +20,17 @@ create table products (
 );
 
 create table sale (
-`id` int primary key auto_increment,
+`id` bigint primary key auto_increment,
 `high_date` date not null,
 `total` double,
-`id_client` int not null,
+`id_client` bigint not null,
 constraint `fk_id_client` foreign key (`id_client`) references `clients` (`id`)
 );
 
 create table sale_detail(
-`id` int primary key auto_increment,
-`id_sale` int not null,
-`id_product` int not null,
+`id` bigint primary key auto_increment,
+`id_sale` bigint not null,
+`id_product` bigint not null,
 `quantity` int not null,
 `subtotal` double not null,
 constraint `fk_id_sale` foreign key (`id_sale`) references `sale` (`id`),
